requirements.txt
appdirs==1.4.3
click==6.7
cssmin==0.2.0
Flask==0.12.1
Flask-Assets==0.12
Flask-Login==0.4.0
Flask-SQLAlchemy==2.2
itsdangerous==0.24
Jinja2==2.9.6
jsmin==2.2.2
MarkupSafe==1.0
packaging==16.8
pyparsing==2.2.0
six==1.10.0
SQLAlchemy==1.1.9
webassets==0.12.1
Werkzeug==0.12.1
psycopg2==2.7.7
Pillow==6.0.0
boto3==1.9.138
pytest==4.4
selenium==3.141.0
